# Optimze VDF