--- a/fastcrypto-vdf/benches/class_group.rs
+++ b/fastcrypto-vdf/benches/class_group.rs
@@ -3,11 +3,11 @@
 
 use criterion::measurement::Measurement;
 use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkGroup, Criterion};
+use fastcrypto::groups::Doubling;
 use fastcrypto_vdf::class_group::{Discriminant, QuadraticForm};
 use fastcrypto_vdf::ParameterizedGroupElement;
 use num_bigint::BigInt;
 use num_traits::Num;
-use fastcrypto::groups::Double;
 
 fn class_group_ops_single<M: Measurement>(
     discriminant_string: &str,