--- a/fastcrypto-vdf/benches/vdf.rs
+++ b/fastcrypto-vdf/benches/vdf.rs
@@ -6,15 +6,15 @@ extern crate criterion;
 
 use criterion::measurement::Measurement;
 use criterion::{BenchmarkGroup, BenchmarkId, Criterion};
+use fastcrypto::groups::multiplier::windowed::WindowedScalarMultiplier;
 use fastcrypto_vdf::class_group::{Discriminant, QuadraticForm};
+use fastcrypto_vdf::hash_prime::DefaultPrimalityCheck;
 use fastcrypto_vdf::vdf::wesolowski::{FastVerify, StrongFiatShamir, StrongVDF};
 use fastcrypto_vdf::vdf::VDF;
 use fastcrypto_vdf::Parameter;
 use num_bigint::BigInt;
 use num_traits::Num;
 use rand::{thread_rng, RngCore};
-use fastcrypto::groups::multiplier::windowed::WindowedScalarMultiplier;
-use fastcrypto_vdf::hash_prime::DefaultPrimalityCheck;
 
 struct VerificationInputs {
     iterations: u64,
@@ -46,12 +46,14 @@ fn verify_single<M: Measurement>(parameters: VerificationInputs, c: &mut Benchma
         b.iter(|| vdf.verify(&input, &result, &proof))
     });
 
-    let fast_verify: FastVerify<QuadraticForm, StrongFiatShamir<QuadraticForm, 264, DefaultPrimalityCheck>,
-        WindowedScalarMultiplier<QuadraticForm, BigInt, 256, 34, 5>> = FastVerify::new(vdf2, input2);
+    let fast_verify: FastVerify<
+        QuadraticForm,
+        StrongFiatShamir<QuadraticForm, 264, DefaultPrimalityCheck>,
+        WindowedScalarMultiplier<QuadraticForm, BigInt, 256, 34, 5>,
+    > = FastVerify::new(vdf2, input2);
     c.bench_function(format!("{} fast", discriminant_size), move |b| {
         b.iter(|| fast_verify.fast_verify(&result2, &proof2))
     });
-
 }
 
 fn verify(c: &mut Criterion) {