Various optimisations on VDF verification. The most important is using multi-scalar multiplication.

VDF verification is optimized ~55%:

| Discriminant size     | Old       | New       |
|------|-----------|-----------|
| 1024 | 72.343 ms | 32.998 ms |
| 2048 | 167.66 ms | 79.427 ms |